backend/requirements.txt
# Backend Python dependencies
# Install with: pip install -r requirements.txt

# ── Core Framework ────────────────────────────────
fastapi==0.111.0           # FastAPI web framework
uvicorn[standard]==0.30.1  # ASGI server for development; Gunicorn uses it in production

# ── Database ──────────────────────────────────────
sqlalchemy==2.0.30         # ORM for database models
# SQLite is built into Python - no extra driver needed for local dev.
# For MySQL (production) uncomment:
# pymysql==1.1.1           # MySQL driver for SQLAlchemy
# cryptography==42.0.8     # Required by PyMySQL for SSL

# ── Environment & Config ──────────────────────────
python-dotenv==1.0.1       # Load environment variables from .env file

# ── Data Validation ───────────────────────────────
pydantic==2.7.1            # Used internally by FastAPI for request/response schemas

# ── Authentication ────────────────────────────────
python-jose[cryptography]==3.3.0  # JWT token creation and verification
passlib[bcrypt]==1.7.4            # Password hashing (bcrypt)
bcrypt==4.0.1                     # passlib requires bcrypt <5.0

# ── Production Server ─────────────────────────────
gunicorn==22.0.0           # WSGI/ASGI process manager for production deployment

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi>=0.1.9            # IP-based rate limiting for FastAPI (no Redis required)

# ── Static File Serving ────────────────────────────────────────────────────────
aiofiles>=23.2.0          # Required by starlette.staticfiles.StaticFiles (admin SPA)

# ── Migrations ────────────────────────────────────────────────────────────────
alembic>=1.13.0           # Database schema migrations (version-controlled)

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=9.0.0             # Test runner
httpx>=0.27.0             # ASGI test client (httpx.TestClient)
